requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
praw==7.7.1
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1